scripts/requirements-detector.txt
numpy
onnxruntime
pillow
huggingface_hub
